+++ aiohomematic/model/update.py
@@ -96,6 +96,10 @@ class DpUpdate(CallbackDataPoint, PayloadMixin):
             return self._device.available_firmware
         return self._device.firmware
 
+    def _get_signature(self) -> str:
+        """Return the signature of the data_point."""
+        return f"{self._category}/{self._device.model}"
+
     def _get_path_data(self) -> DataPointPathData:
         """Return the path data of the data_point."""
         return DataPointPathData(