+++ changelog.md
@@ -1,6 +1,8 @@
 # Version 2025.8.9 (2025-08-23)
 
+- Add signature to model
 - Improve immutability
+- Improve readability of visibility cache
 
 # Version 2025.8.8 (2025-08-23)